+++ UICommon/_events/EventUtils.ts
@@ -1,6 +1,9 @@
 /// <amd-module name="UICommon/_events/EventUtils" />
 /* tslint:disable */
 
+import {Record} from 'Types/entity';
+import {Logger} from 'UICommon/Utils';
+
 /**
  * @author Тэн В.А.
  */
@@ -125,4 +128,46 @@ export function keysHandler(event, keys, handlerSet, scope: object, dontStop: bo
          }
       }
    }
+}
+
+export function checkBindValue(event, value): boolean {
+   const checkNested = (obj, propName, index) => {
+      if (obj === undefined) {
+         return false;
+      }
+      if (obj === null) {
+         return true;
+      }
+      if (propName[index] in obj && propName.length === index + 1) {
+         return true;
+      }
+      if (Array.isArray(obj[propName[index]])) {
+         // могли сделать bind на массив внутри объекта, надо проверить что все поля совпадают
+         let checkArray = [];
+         for (let i = 0; i < obj[propName[index]].length; i++) {
+            checkArray.push(checkNested(obj[propName[index]][i], propName, index + 1));
+         }
+         return checkArray.indexOf(false) <= -1;
+      }
+      if (obj[propName[index]] instanceof Record) {
+         return true
+      }
+      return checkNested(obj[propName[index]], propName, index + 1);
+   };
+
+   if (!value) {
+      return false;
+   }
+   const data = event.data;
+   const valueArray = value.split('.');
+   const re = /\[\s*\S*]/ig;
+   for (const i in valueArray) {
+      if (re.test(valueArray[i])) {
+         valueArray[i] = valueArray[i].split(re)[0];
+      }
+   }
+   if (!checkNested(data, valueArray, 0)) {
+      Logger.warn(`Bind на несуществующее поле "${value}". Bind может работать не правильно`, event.viewController);
+   }
+   return true;
 }
