+++ UIInferno/UICore/_executor/_Markup/Generator.ts
@@ -496,52 +496,12 @@ export class Generator {
          return checkResult.call(this, res, type, name);
       }
    };
-   private static checkBindValue(event, value) {
-      const checkNested = (obj, propName, index) => {
-         if (obj === undefined) {
-            return false;
-         }
-         if (obj === null) {
-            return true;
-         }
-         if (propName[index] in obj && propName.length === index + 1) {
-            return true;
-         }
-         if (Array.isArray(obj[propName[index]])) {
-            // могли сделать bind на массив внутри объекта, надо проверить что все поля совпадают
-            let checkArray = [];
-            for (let i = 0; i < obj[propName[index]].length; i++) {
-               checkArray.push(checkNested(obj[propName[index]][i], propName, index + 1));
-            }
-            return checkArray.indexOf(false) <= -1;
-         }
-         if (obj[propName[index]] instanceof Record) {
-            return true
-         }
-         return checkNested(obj[propName[index]], propName, index + 1);
-      };
 
-      if (!value) {
-         return false;
-      }
-      const data = event.data;
-      const valueArray = value.split('.');
-      const re = /\[\s*\S*]/ig;
-      for (const index in valueArray) {
-         if (re.test(valueArray[index])) {
-            valueArray[index] = valueArray[index].split(re)[0];
-         }
-      }
-      if (!checkNested(data, valueArray, 0)) {
-         Logger.warn(`Bind на несуществующее поле "${value}". Bind может работать не правильно`, event.viewController);
-      }
-      return true;
-   }
    prepareEvents(events) {
       Object.keys(events).forEach((eventName) => {
          const eventArr = events[eventName];
          eventArr.forEach((event) => {
-            if (Generator.checkBindValue(event, event.bindValue)) {
+            if (EventUtils.checkBindValue(event, event.bindValue)) {
                event.fn = function (eventObj, value) {
                   if (!event.handler(this.viewController, value)) {
                      event.handler(this.data, value);
